+++ src/com/awesome/excelpp/math/Count.java
@@ -1,8 +1,8 @@
 package com.awesome.excelpp.math;
 
 public class Count extends Formula {
-	public int getValue(int...numbers) {
-		int count = 0;
+	public double getValue(double...numbers) {
+		double count = 0;
 		for(int i=0; i < numbers.length; i++) 
 			count = i;
 		return count;
