+++ src/com/awesome/excelpp/math/Min.java
@@ -1,8 +1,8 @@
 package com.awesome.excelpp.math;
 
 public class Min extends Formula {
-	public int getValue(int...numbers) {
-		int minimum = 0;
+	public double getValue(double...numbers) {
+		double minimum = 0;
 		for(int i=0; i<numbers.length; i++)
 			if( minimum < numbers[i]) 
 				minimum = numbers[i];
